init of trackers consents moved to _app
