+++ .storybook/main.js
@@ -4,7 +4,7 @@ module.exports = {
   stories: [
     '../theme/*.stories.tsx',
     '../features/**/*.stories.tsx',
-    '../components/**/*.stories.tsx ',
+    '../components/**/*.stories.tsx',
   ],
   addons: ['@storybook/addon-links', '@storybook/addon-essentials', '@storybook/addon-postcss'],
   // typescript: {
